moved folders for sessionmanager
