+++ aux_ui/lib/generic_classes/session_manager.dart
